use fd as default fzf command
This introduces `fd` as a dependency of `.zshrc`

`fd` claims to be a faster and more usable replacement for `find`. By
default it also ignores patterns from .gitignore and hidden directories
or files.

`ripgrep` can also search file names, but I didn't look much into the
pros/cons of using it instead.

Resources:
https://github.com/junegunn/fzf#respecting-gitignore
https://github.com/sharkdp/fd#using-fd-with-fzf