don't use a bare repository anymore
A bare repository doesn't really fit the use case here because I don't
intend for this to live on a server - I still want to make and add
changes.

Instead I change the default working tree for this repository.

A benefit of this is by defualt commands like `git status` will let me
know if I'm ahead/behind the upstream. This could probably be fixed by
updating the config option `remote.origin.fetch` refspec to be correct,
but I'd rather just get it by defualt.

Another benefit is that if I'm in the `~/dotfiles` directory I can use
the normal git CLI (and my git aliases) and not need my `dot` alias.

Resources:
https://git-scm.com/docs/git-config#Documentation/git-config.txt-corebare
https://git-scm.com/docs/git-config#Documentation/git-config.txt-remoteltnamegtfetch
https://git-scm.com/docs/git-fetch#Documentation/git-fetch.txt-ltrefspecgt